pyproject.toml
[tool.poetry]
name = "astrotools"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas="^2.1.4"
pyswisseph="2.10.3.2"
pytz="2023.3.post1"
Requests="2.31.0"
scipy="1.11.4"
swisseph="0.0.0.dev1"
timezonefinder="6.2.0"
pyhumps = "^3.8.0"
pytest = "^7.4.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
